Rename some packages and move some classes
