+++ deployment/src/test/java/com/datastax/oss/quarkus/CassandraClientConfigTest.java
@@ -21,7 +21,7 @@ import static com.datastax.oss.driver.api.core.config.DefaultDriverOption.REQUES
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.datastax.oss.driver.api.core.config.DriverExecutionProfile;
-import com.datastax.oss.quarkus.runtime.api.driver.QuarkusCqlSession;
+import com.datastax.oss.quarkus.runtime.api.session.QuarkusCqlSession;
 import io.quarkus.test.QuarkusUnitTest;
 import io.quarkus.test.common.QuarkusTestResource;
 import java.time.Duration;
