+++ runtime/src/main/java/com/datastax/oss/quarkus/config/CassandraClientConnectionConfig.java
@@ -15,7 +15,7 @@
  */
 package com.datastax.oss.quarkus.config;
 
-import com.datastax.oss.quarkus.runtime.api.driver.QuarkusCqlSession;
+import com.datastax.oss.quarkus.runtime.api.session.QuarkusCqlSession;
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
 import java.time.Duration;
