+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/CassandraClientRecorder.java
@@ -16,7 +16,7 @@
 package com.datastax.oss.quarkus.runtime;
 
 import com.datastax.oss.quarkus.config.CassandraClientConfig;
-import com.datastax.oss.quarkus.runtime.api.driver.QuarkusCqlSession;
+import com.datastax.oss.quarkus.runtime.api.session.QuarkusCqlSession;
 import com.datastax.oss.quarkus.runtime.metrics.MetricsConfig;
 import com.datastax.oss.quarkus.runtime.metrics.NoopMetricRegistry;
 import io.quarkus.arc.Arc;
