+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/api/reactive/MutinyReactiveResultSet.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.quarkus.runtime.api.driver;
+package com.datastax.oss.quarkus.runtime.api.reactive;
 
-import com.datastax.dse.driver.api.core.cql.continuous.reactive.ContinuousReactiveResultSet;
+import com.datastax.dse.driver.api.core.cql.reactive.ReactiveResultSet;
 import com.datastax.dse.driver.api.core.cql.reactive.ReactiveRow;
 import io.smallrye.mutiny.Multi;
 
-public interface MutinyContinuousReactiveResultSet
-    extends Multi<ReactiveRow>, ContinuousReactiveResultSet, MutinyReactiveQueryMetadata {}
+public interface MutinyReactiveResultSet
+    extends Multi<ReactiveRow>, ReactiveResultSet, MutinyReactiveQueryMetadata {}
