+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/api/session/QuarkusCqlSessionBuilder.java
@@ -13,23 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.quarkus.runtime.internal;
+package com.datastax.oss.quarkus.runtime.api.session;
 
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.datastax.oss.driver.api.core.config.DriverConfigLoader;
 import com.datastax.oss.driver.api.core.context.DriverContext;
 import com.datastax.oss.driver.api.core.session.ProgrammaticArguments;
 import com.datastax.oss.driver.api.core.session.SessionBuilder;
-import com.datastax.oss.quarkus.runtime.api.driver.QuarkusCqlSession;
+import com.datastax.oss.quarkus.runtime.internal.context.QuarkusDriverContext;
+import com.datastax.oss.quarkus.runtime.internal.reactive.DefaultQuarkusCqlSession;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import org.eclipse.microprofile.metrics.MetricRegistry;
 
-public class QuarkusSessionBuilder
-    extends SessionBuilder<QuarkusSessionBuilder, QuarkusCqlSession> {
+public class QuarkusCqlSessionBuilder
+    extends SessionBuilder<QuarkusCqlSessionBuilder, QuarkusCqlSession> {
 
   private final MetricRegistry metricRegistry;
 
-  public QuarkusSessionBuilder(MetricRegistry metricRegistry) {
+  public QuarkusCqlSessionBuilder(@NonNull MetricRegistry metricRegistry) {
     this.metricRegistry = metricRegistry;
   }
