+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/internal/context/QuarkusDriverContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.quarkus.runtime.internal;
+package com.datastax.oss.quarkus.runtime.internal.context;
 
 import com.datastax.oss.driver.api.core.config.DriverConfigLoader;
 import com.datastax.oss.driver.api.core.session.ProgrammaticArguments;
