+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/internal/reactive/Wrappers.java
@@ -13,22 +13,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.quarkus.runtime.reactive;
+package com.datastax.oss.quarkus.runtime.internal.reactive;
 
+import edu.umd.cs.findbugs.annotations.NonNull;
 import io.smallrye.mutiny.Multi;
 import io.vertx.core.Context;
 import io.vertx.core.Vertx;
+import java.util.concurrent.Executor;
 import org.reactivestreams.Publisher;
 
 public class Wrappers {
+
   public static <T> Multi<T> toMulti(Publisher<T> publisher) {
+    Multi<T> multi = Multi.createFrom().publisher(publisher);
     Context context = Vertx.currentContext();
     if (context != null) {
-      return Multi.createFrom()
-          .publisher(publisher)
-          .emitOn(command -> context.runOnContext(x -> command.run()));
-    } else {
-      return Multi.createFrom().publisher(publisher);
+      multi = multi.emitOn(new VertexContextExecutor(context));
+    }
+    return multi;
+  }
+
+  private static class VertexContextExecutor implements Executor {
+
+    private final Context context;
+
+    public VertexContextExecutor(Context context) {
+      this.context = context;
+    }
+
+    @Override
+    public void execute(@NonNull Runnable command) {
+      context.runOnContext(x -> command.run());
     }
   }
 }
